+++ src/app/index.run.js
@@ -1,9 +1,7 @@
 import { Visualizer } from '@uirouter/visualizer';
 import { states as dashboardStates } from './pages/dashboard/dashboard.state.js';
 import { states as listingStates } from './pages/listing/listing.state.js';
-import { states as organizationsStates } from './pages/organizations/organizations.state.js';
 import { states as surveillanceStates } from './pages/surveillance/surveillance.state.js';
-import { states as usersStates } from './pages/users/users.state.js';
 
 (() => {
     'use strict';
@@ -46,22 +44,13 @@ import { states as usersStates } from './pages/users/users.state.js';
                             $uiRouter.stateRegistry.register(state);
                             needsReload = needsReload || $state.$current.name === state.name;
                         });
-                    }
-
-                    if (featureFlags.isOn('organizations')) {
-                        organizationsStates['enabled'].forEach(state => {
+                    } else {
+                        listingStates['listing-edit-off'].forEach(state => {
                             if ($uiRouter.stateRegistry.get(state.name)) {
                                 $uiRouter.stateRegistry.deregister(state.name);
                             }
                             $uiRouter.stateRegistry.register(state);
                             needsReload = needsReload || $state.$current.name === state.name;
-                        });
-                    } else {
-                        organizationsStates['enabled'].forEach(state => {
-                            if ($uiRouter.stateRegistry.get(state.name)) {
-                                $uiRouter.stateRegistry.deregister(state.name);
-                            }
-                            needsRedirect = needsRedirect || $state.$current.name === state.name;
                         });
                     }
 
@@ -99,16 +88,6 @@ import { states as usersStates } from './pages/users/users.state.js';
                         });
                     }
 
-                    if (featureFlags.isOn('ocd2749')) {
-                        usersStates['ocd2749-on'].forEach(state => {
-                            if ($uiRouter.stateRegistry.get(state.name)) {
-                                $uiRouter.stateRegistry.deregister(state.name);
-                            }
-                            $uiRouter.stateRegistry.register(state);
-                            needsReload = needsReload || $state.$current.name === state.name;
-                        });
-                    }
-
                     // Display ui-router state changes
                     if (featureFlags.isOn('states')) {
                         $uiRouter.plugin(Visualizer);
@@ -123,7 +102,7 @@ import { states as usersStates } from './pages/users/users.state.js';
                 });
         };
 
-        if (authService.hasAnyRole()) {
+        if (authService.hasAnyRole(['ROLE_ADMIN', 'ROLE_ONC', 'ROLE_ACB', 'ROLE_ATL', 'ROLE_CMS_STAFF', 'ROLE_DEVELOPER'])) {
             networkService.keepalive()
                 .then(() => {
                     loadFlags();