+++ src/app/pages/dashboard/dashboard.component.js
@@ -43,8 +43,8 @@ export const DashboardComponent = {
             let that = this;
             switch (action) {
             case 'delete':
-                this.networkService.removeUserFromDeveloper(data)
-                    .then(() => that.networkService.getUsersAtDeveloper(this.developerId).then(response => that.users = response.users));
+                this.networkService.removeUserFromDeveloper(data, this.developerId)
+                    .then(() => that.networkService.getUsersAtDeveloper(that.developerId).then(response => that.users = response.users));
                 break;
             case 'invite':
                 this.networkService.inviteUser({
@@ -61,7 +61,7 @@ export const DashboardComponent = {
                 this.networkService.getUsersAtDeveloper(this.developerId)
                     .then(response => that.users = response.users);
                 break;
-            case 'reload':
+            case 'impersonate':
                 this.$state.reload();
                 break;
                 //no default