+++ RELEASE_NOTES.md
@@ -1,5 +1,56 @@
 # Release Notes
 
+## Version 10.26.0
+_23 September 2019_
+
+### Features
+* Add new "Organizations" navigation item for logged in users
+  * Complete change / upgrade of ONC-ACB and ONC-ATL management screens
+  * Add new Organizations - Developer page
+  * Includes support for Edit, Merge, Split
+  * Add support for Product editing on new Organizations - Developer page
+  * Includes Edit, Merge, and Split
+  * Add support for Version editing on new Organizations - Developer - Product page
+  * Includes Edit, Merge, and Split
+
+### Flagged Feature
+* Allow ONC-ACBs to generate their required quarterly and annual Surveillance Reports
+
+### User Interface Updates
+* Standardize button display on complaints
+* Improve display of User Organizations
+
+---
+
+## Version 10.25.0
+_10 September 2019_
+
+### Features
+* Update user information to include
+  * Last Logged In Date
+  * Organization(s) user is associated with
+* Update the API Key Usage report UI to be consistent with other reports
+* Provide new UI/UX for user management
+
+### Display changes
+* Use correct title for ONC-ATL
+
+---
+
+## Version 10.24.1
+_26 August 2019_
+
+### User Interface Updates
+* Change title of Complaints page
+* Turn off EDGE's telephone number detection "feature"
+
+### Bug Fixes
+* Limit access to pages / features with explicit ROLE lists
+* Update surveillance filter to work correctly when navigating back to search
+* Use correct link on Manage Surveillance page based on active flag
+
+---
+
 ## Version 10.24.0
 _12 August 2019_
 
@@ -17,10 +68,10 @@ _12 August 2019_
 _29 July 2019_
 
 ### Flagged Features
-* update administrator's navigation ability
+* Complete revamp of administration navigation
+* Add "Save Search" functionality to admin reports
 
 ---
-
 
 ## Version 10.22.0
 _22 July 2019_